requirements.txt
langchain
langgraph
openai
google-generativeai
duckduckgo-search
beautifulsoup4
requests
python-dotenv
